move the Server class out of bentos.py to its own module
