+++ src/bentoml/bentos.py
@@ -20,10 +20,9 @@ from ._internal.bento import Bento
 from ._internal.utils import resolve_user_filepath
 from ._internal.bento.build_config import BentoBuildConfig
 from ._internal.configuration.containers import BentoMLContainer
+from ._internal.server.server import Server
 
 if TYPE_CHECKING:
-    from bentoml.client import Client
-
     from ._internal.bento import BentoStore
 
 logger = logging.getLogger(__name__)
@@ -499,27 +498,3 @@ def serve(
     process = subprocess.Popen(args)
 
     return Server(process, host, port)
-
-
-class Server:
-    def __init__(self, process: subprocess.Popen[bytes], host: str, port: int) -> None:
-        self._process = process
-        self._host = host
-        self._port = port
-
-    def get_client(self) -> Client:
-        from bentoml.client import Client
-
-        Client.wait_until_server_is_ready(self._host, self._port, 10)
-        return Client.from_url(f"http://localhost:{self._port}")
-
-    def stop(self) -> None:
-        self.process.kill()
-
-    @property
-    def process(self) -> subprocess.Popen[bytes]:
-        return self._process
-
-    @property
-    def address(self) -> str:
-        return f"{self._host}:{self._port}"
